Fix: fps dinámico según duración del video + config panel defensivo
- CloudConvertSpinService: tipo fps int→float para precisión decimal
- SceneController: captura video_duration del frontend, calcula fps=72/duración
- SpinController: mismo cálculo dinámico de fps
- dataScene: hidden input video_duration + captura via loadedmetadata
- config.blade.php: try/catch + data_get para evitar 500 si migración no existe
- Properties: accessor active_spin defensivo con try/catch
- Fallback fps=1.8 (~40s video) si no se recibe duración

+++ app/Services/CloudConvertSpinService.php
@@ -24,7 +24,7 @@ class CloudConvertSpinService
      * 3) archive -> zip de todos los frames
      * 4) export/url -> URL del zip
      */
-    public function createSpinJob(int $fps = 2, string $imgExt = 'jpg'): Job
+    public function createSpinJob(float $fps = 2.0, string $imgExt = 'jpg'): Job
     {
         $cloudconvert = $this->client();
         $desiredFrames = 72;
